feat(database): migrate from SQLite to Oracle, updating connection handling and service registrations

+++ relevo-api/src/Relevo.UseCases/Contributors/ContributorDTO.cs
@@ -1,2 +1,2 @@
 ﻿namespace Relevo.UseCases.Contributors;
-public record ContributorDTO(int Id, string Name, string? PhoneNumber);
+public record ContributorDTO(decimal Id, string Name, string? PhoneNumber);
