+++ relevo-api/src/Relevo.Web/Contributors/ContributorRecord.cs
@@ -1,3 +1,3 @@
 ﻿namespace Relevo.Web.Contributors;
 
-public record ContributorRecord(int Id, string Name, string? PhoneNumber);
+public record ContributorRecord(decimal Id, string Name, string? PhoneNumber);
